feat: update schema

+++ packages/server/src/routes/app.ts
@@ -1,37 +1,26 @@
 import { db } from '@/db'
-import { apps, organizations } from '@/db/schema'
+import { apps } from '@/db/schema'
 import type { UserClaims } from '@/types'
-import { generateWebhookUrl } from '@/utils/webhook'
-import { v4 } from '@lukeed/uuid/secure'
 import { and, count, eq, sql } from 'drizzle-orm'
 import { t } from 'elysia'
 import type { APIGroupServerType } from '..'
 
 export async function addAppRoutes(path: string, server: APIGroupServerType) {
-  // get all apps under the organization
+  // get all user's apps
   server.get(
     path,
-    async ({ query, params, bearer, jwt, set }) => {
+    async ({ query, bearer, jwt }) => {
       const user = (await jwt.verify(bearer)) as UserClaims
-      const organization = await db.query.organizations.findFirst({
-        where: and(eq(organizations.id, params.orgId), eq(organizations.creatorId, user.id))
-      })
-      if (!organization) {
-        set.status = 404
-        return 'Organization not found'
-      }
+      const cond = eq(apps.creatorId, user.id)
       const list = await db.query.apps.findMany({
-        where: eq(apps.organizationId, params.orgId),
+        where: cond,
         offset: query.offset ?? 0,
         limit: query.limit ?? 10
       })
-      const total = await db.select({ value: count() }).from(apps).where(eq(apps.organizationId, params.orgId))
+      const total = await db.select({ value: count() }).from(apps).where(cond)
       return { list, total: total[0].value }
     },
     {
-      params: t.Object({
-        orgId: t.String()
-      }),
       query: t.Object({
         offset: t.MaybeEmpty(t.Numeric()),
         limit: t.MaybeEmpty(t.Numeric())
@@ -39,35 +28,49 @@ export async function addAppRoutes(path: string, server: APIGroupServerType) {
     }
   )
 
+  // get a app and its participants
+  server.get(
+    `${path}/:id`,
+    async ({ params, bearer, jwt }) => {
+      const user = (await jwt.verify(bearer)) as UserClaims
+      const app = await db.query.apps.findFirst({
+        with: {
+          invitedUsers: {
+            columns: {
+              id: true,
+              nickname: true,
+              username: true,
+              avatar: true
+            }
+          }
+        },
+        where: and(eq(apps.creatorId, user.id), eq(apps.id, params.id))
+      })
+      return app
+    }
+  )
+
   // create a new app
   server.post(
     path,
-    async ({ body, params, bearer, jwt, set }) => {
+    async ({ body, bearer, jwt, set }) => {
       const user = (await jwt.verify(bearer)) as UserClaims
-      const organization = await db.query.organizations.findFirst({
-        where: and(eq(organizations.id, params.orgId), eq(organizations.creatorId, user.id))
+      // 检查应用名称是否已存在
+      const existingApp = await db.query.apps.findFirst({
+        where: eq(apps.name, body.name)
       })
-      if (!organization) {
-        set.status = 404
-        return 'Organization not found'
-      }
-      const sameApp = await db.query.apps.findFirst({
-        where: and(eq(apps.name, body.name), eq(apps.organizationId, params.orgId))
-      })
-      if (sameApp) {
+
+      if (existingApp) {
         set.status = 400
-        return 'App already exists'
+        return '应用名称已存在'
       }
       try {
-        const id = v4()
         const ret = await db
           .insert(apps)
           .values([
             {
               ...body,
-              id,
-              organizationId: params.orgId,
-              generatedWebhookUrl: generateWebhookUrl(body.upstreamRepoType, params.orgId, id)
+              creatorId: user.id
             }
           ])
           .returning()
@@ -81,16 +84,10 @@ export async function addAppRoutes(path: string, server: APIGroupServerType) {
       }
     },
     {
-      params: t.Object({
-        orgId: t.String()
-      }),
       body: t.Object({
         name: t.String(),
-        upstreamRepoType: t.Enum({ Codeup: 'codeup' }),
-        upstreamRepoUrl: t.MaybeEmpty(t.String()),
-        upstreamSecretToken: t.MaybeEmpty(t.String()),
-        giteaRepo: t.String(),
-        giteaToken: t.MaybeEmpty(t.String()),
+        logo: t.MaybeEmpty(t.String()),
+        description: t.MaybeEmpty(t.String())
       })
     }
   )
@@ -100,19 +97,15 @@ export async function addAppRoutes(path: string, server: APIGroupServerType) {
     `${path}/:id`,
     async ({ body, params, bearer, jwt, set }) => {
       const user = (await jwt.verify(bearer)) as UserClaims
-      const organization = await db.query.organizations.findFirst({
-        where: and(eq(organizations.id, params.orgId), eq(organizations.creatorId, user.id))
-      })
-      if (!organization) {
-        set.status = 404
-        return 'Organization not found'
-      }
       try {
-        const ret = await db.update(apps).set({
-          ...body,
-          generatedWebhookUrl: generateWebhookUrl(body.upstreamRepoType, params.orgId, params.id),
-          updatedAt: sql`(datetime('now', 'localtime'))`
-        }).where(eq(apps.id, params.id)).returning()
+        const ret = await db
+          .update(apps)
+          .set({
+            ...body,
+            updatedAt: sql`(datetime('now', 'localtime'))`
+          })
+          .where(and(eq(apps.creatorId, user.id), eq(apps.id, params.id)))
+          .returning()
         if (ret.length > 0) {
           return ret[0]
         }
@@ -124,41 +117,32 @@ export async function addAppRoutes(path: string, server: APIGroupServerType) {
     },
     {
       params: t.Object({
-        id: t.String(),
-        orgId: t.String()
+        id: t.String()
       }),
       body: t.Object({
-        name: t.String(),
-        upstreamRepoType: t.Enum({ Codeup: 'codeup' }),
-        upstreamRepoUrl: t.MaybeEmpty(t.String()),
-        upstreamSecretToken: t.MaybeEmpty(t.String()),
-        giteaRepo: t.String(),
-        giteaToken: t.MaybeEmpty(t.String()),
+        logo: t.MaybeEmpty(t.String()),
+        description: t.MaybeEmpty(t.String())
       })
     }
   )
 
   // delete an app
-  server.delete(`${path}/:id`, async ({ params, bearer, jwt, set }) => {
-    const user = (await jwt.verify(bearer)) as UserClaims
-    const organization = await db.query.organizations.findFirst({
-      where: and(eq(organizations.id, params.orgId), eq(organizations.creatorId, user.id))
-    })
-    if (!organization) {
-      set.status = 404
-      return 'Organization not found'
+  server.delete(
+    `${path}/:id`,
+    async ({ params, bearer, jwt, set }) => {
+      const user = (await jwt.verify(bearer)) as UserClaims
+      try {
+        const ret = await db.delete(apps).where(and(eq(apps.creatorId, user.id), eq(apps.id, params.id))).returning()
+        return ret.length > 0
+      } catch (error) {
+        set.status = 500
+        return 'Failed to delete app'
+      }
+    },
+    {
+      params: t.Object({
+        id: t.String()
+      })
     }
-    try {
-      const ret = await db.delete(apps).where(eq(apps.id, params.id)).returning()
-      return ret.length > 0
-    } catch (error) {
-      set.status = 500
-      return 'Failed to delete app'
-    }
-  }, {
-    params: t.Object({
-      id: t.String(),
-      orgId: t.String()
-    })
-  })
+  )
 }
